# apps/api/src/hyp3_api/api-spec/openapi-spec.yml
openapi: 3.0.1

info:
  title: hyp3-api
  version: ""

security:
  - EarthDataLogin: []

paths:

  /jobs:

    post:
      description: Submits a list of jobs for processing.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/post_jobs_body"
        required: true
      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/jobs_response"

    get:
      description: Get list of previously run jobs.
      parameters:
        - name: status_code
          in: query
          schema:
            $ref: "#/components/schemas/status_code"
        - name: start
          in: query
          schema:
            $ref: "#/components/schemas/datetime"
        - name: end
          in: query
          schema:
            $ref: "#/components/schemas/datetime"
        - name: name
          in: query
          schema:
            $ref: "#/components/schemas/name"
        - name: job_type
          in: query
          schema:
            $ref: "./job_parameters.yml#components/schemas/job_type"
        - name: start_token
          in: query
          schema:
            $ref: "#/components/schemas/start_token"

      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/jobs_response"

  /jobs/{job_id}:
    get:
      description: Get a previously run job.
      parameters:
        - name: job_id
          in: path
          schema:
            $ref: "#/components/schemas/job_or_subscription_id"
          required: true

      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/job"

  /user:
    get:
      description: Get information about the logged in user.
      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"

  /subscriptions:
    post:
      description: Adds a subscription with seach criteria and processing parameters to use to process data as it becomes available
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/subscriptions_body"
        required: true
      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/subscriptions_response"
    get:
      description: Get information about subscriptions for the logged in user
      responses:
        "200":
          description: 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/subscriptions_response"

components:
  schemas:

    post_jobs_body:
      description: List for new jobs to submit for processing.
      type: object
      required:
        - jobs
      additionalProperties: false
      properties:
        validate_only:
          $ref: "#/components/schemas/validate_only"
        jobs:
          $ref: "#/components/schemas/list_of_new_jobs"

    jobs_response:
      description: List of submitted jobs.
      type: object
      required:
        - jobs
      additionalProperties: false
      properties:
        validate_only:
          $ref: "#/components/schemas/validate_only"
        jobs:
          $ref: "#/components/schemas/list_of_jobs"
        next:
          $ref: "#/components/schemas/next_url"

    subscriptions_body:
      description: foo
      type: object
      required:
        - search_parameters
        - job_specification
      properties:
        search_parameters:
          $ref: "#/components/schemas/search_parameters"
        job_specification:
          $ref: "#components/schemas/subscription_job_specification"

    subscriptions_response:
      description: foo
      type: object
      required:
        - subscription_id
        - search_parameters
        - job_specification
      properties:
        subscription_id:
          $ref: "#/components/schemas/job_or_subscription_id"
        search_parameters:
          $ref: "#/components/schemas/search_parameters"
        job_specification:
          $ref: "#components/schemas/subscription_job_specification"


    user:
      description: Information about a user (quota, user id)
      type: object
      required:
        - user_id
        - quota
      additionalProperties: false
      properties:
        user_id:
          $ref: "#/components/schemas/user_id"
        quota:
          $ref: "#/components/schemas/quota"
        job_names:
          $ref: "#components/schemas/job_names_list"

    quota:
      description: Containes the limit of jobs per month and the amount remaining for a user.
      type: object
      required:
        - max_jobs_per_month
        - remaining
      additionalProperties: false
      properties:
        max_jobs_per_month:
          type: integer
          minimum: 0
        remaining:
          type: integer
          minimum: 0

    job_names_list:
      type: array
      items:
        $ref: "#/components/schemas/name"

    list_of_new_jobs:
      description: Contains a list of new job objects.
      type: array
      minItems: 1
      maxItems: 200
      items:
        $ref: "#/components/schemas/new_job"

    subscription_job_specification:
      description: Contains user provided information on running a job for a subscription.
      type: object
      required:
        - job_type
        - name
      properties:
        job_type:
          $ref: "./job_parameters.yml#components/schemas/job_type"
        job_parameters:
          $ref: "./job_parameters.yml#components/schemas/job_parameters"
        name:
          $ref: "#/components/schemas/name"

    new_job:
      description: Contains user provided information on runnning a new job.
      type: object
      required:
        - job_type
        - job_parameters
      additionalProperties: false
      example:
        job_parameters:
          granules:
            - S1A_IW_SLC__1SSV_20150621T120220_20150621T120232_006471_008934_72D8
        job_type: RTC_GAMMA
        name: Job Name
      properties:
        job_type:
          $ref: "./job_parameters.yml#components/schemas/job_type"
        job_parameters:
          $ref: "./job_parameters.yml#components/schemas/job_parameters"
        name:
          $ref: "#/components/schemas/name"


    list_of_jobs:
      type: array
      items:
        $ref: "#/components/schemas/job"

    search_parameters:
      description: parameters used to find data from the ASF Search API
      additionalProperties: true
      type: object
      required:
        - start
        - end
      properties:
        asfframe:
          type: integer
          minimum: 0
          maximum: 1184
        beamMode:
          type: string
          enum:
            - IW
        flightDirection:
          type: string
          enum:
            - ASCENDING
            - DECENDING
        intersectsWith:
          type: string
          format: wkt
        processingLevel:
          type: string
          enum:
            - GRD_HD
            - SLC
        polorization:
          type: string
          enum:
            - VV
            - VV+VH
            - HH
            - HH+HV
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time

    job:
      description: Contains information about a submitted job.
      type: object
      required:
        - job_id
        - user_id
        - job_type
        - job_parameters
        - request_time
        - status_code
      additionalProperties: false
      properties:
        job_id:
          $ref: "#/components/schemas/job_or_subscription_id"
        user_id:
          $ref: "#/components/schemas/user_id"
        job_type:
          $ref: "./job_parameters.yml#components/schemas/job_type"
        job_parameters:
          $ref: "./job_parameters.yml#components/schemas/job_parameters"
        request_time:
          $ref: "#/components/schemas/datetime"
        status_code:
          $ref: "#/components/schemas/status_code"
        name:
          $ref: "#/components/schemas/name"
        files:
          $ref: "#/components/schemas/list_of_files"
        browse_images:
          $ref: "#/components/schemas/list_of_urls"
        thumbnail_images:
          $ref: "#/components/schemas/list_of_urls"
        logs:
          $ref: "#/components/schemas/list_of_urls"
        expiration_time:
          $ref: "#/components/schemas/datetime"

    validate_only:
      type: boolean
      default: false

    job_or_subscription_id:
      description: Unique identifier for a job or subscription
      type: string
      format: uuid
      example: 27836b79-e5b2-4d8f-932f-659724ea02c3

    user_id:
      description: Username from Earthdata Login.
      type: string
      example: myUserId

    datetime:
      description: Date and time object formatted according to ISO 8601
      type: string
      format: date-time
      example: 2020-06-04T18:00:03+00:00

    status_code:
      description: Status of a submitted job.
      type: string
      enum:
        - PENDING
        - RUNNING
        - SUCCEEDED
        - FAILED
      example: SUCCEEDED

    name:
      description: User provided text to name the job
      type: string
      minLength: 1
      maxLength: 20
      example: Job Name

    start_token:
      description: Token used for fetching subsequent results for large queries
      type: string

    next_url:
      description: Url provided for large search results that have been truncated. Use to fetch subsequent results.
      type: string

    list_of_files:
      description: List of downloadable files generated by the job.
      type: array
      items:
        type: object
        required:
          - filename
          - size
          - url
          - s3
        additionalProperties: False
        properties:
          filename:
            type: string
          size:
            type: integer
            minimum: 0
          url:
            type: string
          s3:
            $ref: "#/components/schemas/s3_object"

    s3_object:
      type: object
      additionalProperties: false
      required:
        - bucket
        - key
      properties:
        bucket:
          type: string
        key:
          type: string

    list_of_urls:
      type: array
      items:
        type: string

  securitySchemes:
    EarthDataLogin:
      description: |-
        Authentication requires the user to have an account at urs.earthdata.nasa.gov and log in at auth.asf.alaska.edu
      type: apiKey
      in: cookie
      name: asf-urs